// src/app/model/Book.ts
export class Book {
    id: number;
    download_count: number;
    name: String;
    author_name: String;
    publisher_name: String;
    contact_number: String;
    account_enabled: Boolean;
    fileName: String;
    path: String;
    notes: String;
    type: String;
    file: BinaryType;
}